microservices/search-profile-manager/requirements.txt
flask-lambda-python36
pytest==5.4.1
MarkupSafe==2.0.1
